requirements.txt
numpy>=1.24
sounddevice>=0.4.6
matplotlib>=3.7
paho-mqtt>=1.6.1
PyYAML>=6.0
scipy>=1.10
